Remove hooks directory creation from `repo` helper
This helper previously created a `.git/hooks` directory for every
repository. This wasn't necessary since we only needed this for a few
tests. Fix those tests to create the hooks directory instead.

Reviewed-on: http://gerrit.causes.com/45272
Tested-by: jenkins <jenkins@brigade.com>

+++ lib/overcommit/installer.rb
@@ -119,6 +119,8 @@ module Overcommit
     end
 
     def uninstall_hook_symlinks
+      return unless File.directory?(hooks_path)
+
       Dir.chdir(hooks_path) do
         Overcommit::Utils.supported_hook_types.each do |hook_type|
           FileUtils.rm_rf(hook_type) if overcommit_hook?(hook_type)
